Fixed errors, divided Location and Tower style (you had bugs there), refactored Tower component.
