+++ src/components/Main/Sidebar/Location/Location.jsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import styles from './Location.module.css';
-import TowerContainer from './Tower/TowerContainer';
+import React from "react";
+import styles from "./Location.module.css";
+import { TowersContainer } from "./Towers/TowersContainer";
 
 const Location = (props) => {
   return (
@@ -11,12 +11,12 @@ const Location = (props) => {
         </div>
         <div className={styles.titleWrapper}>
           <div className={styles.title}>Add Location</div>
-          <p className={styles.text}>Did you know you can add favourite locations?</p>
+          <p className={styles.text}>
+            Did you know you can add favourite locations?
+          </p>
         </div>
       </div>
-      <ul className={styles.list}>
-        <TowerContainer />
-      </ul>
+      <TowersContainer />
     </div>
   );
 };
