--- src/components/Main/Sidebar/Location/Tower/Tower.module.css
@@ -1,114 +1,0 @@
-.add {
-  border-radius: 20px 20px 0px 0px;
-  background: var(--main-color);
-  display: flex;
-  align-items: center;
-  padding: 25px 30px 29px;
-  transition: var(--transition);
-  cursor: pointer;
-  border-bottom: 8px solid #f3f3f3;
-}
-
-.add:hover {
-  background: #e17777;
-}
-
-.add:hover .plus {
-  transform: rotate(360deg);
-}
-
-.add:hover .plus {
-  color: #e17777;
-}
-
-.plus {
-  position: relative;
-  flex-shrink: 0;
-  width: 50px;
-  height: 50px;
-  background: #fff;
-  border-radius: 100%;
-  color: var(--main-color);
-  font-size: 35px;
-  font-weight: bold;
-  transition: var(--transition);
-}
-
-.plus:hover {
-  transform: rotate(360deg);
-}
-
-.plus span {
-  position: absolute;
-  top: 42%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  /* display: block; */
-  /* margin-top: -1px;
-  margin-left: -1px; */
-  line-height: 100%;
-  transition: var(--transition);
-}
-
-.titleWrapper {
-  padding-left: 20px;
-  text-align: left;
-}
-
-.title {
-  font-size: 30px;
-}
-
-.text {
-  font-size: 19px;
-}
-
-.list {
-  background: #fff;
-  color: #000;
-  border-radius: 0px 0px 20px 20px;
-  padding: 20px 37px;
-}
-
-.list li {
-  position: relative;
-  text-align: left;
-  padding: 11px 32px 11px 0;
-  border-bottom: 1px solid #f3f3f3;
-  cursor: pointer;
-}
-
-.list li:after {
-  content: '';
-  height: 100%;
-  width: 100%;
-  width: 0;
-  background: #e17777;
-  position: absolute;
-  top: 0;
-  left: 0;
-  transition: var(--transition);
-  z-index: 1;
-}
-
-.list li:hover:after {
-  width: 100%;
-}
-
-.list li:hover a {
-  color: #fff;
-}
-
-.list li a {
-  position: relative;
-  text-transform: uppercase;
-  z-index: 2;
-}
-
-.arrow {
-  display: inline-block;
-  width: 10px;
-  height: 10px;
-  margin-right: 32px;
-  font-size: 12px;
-}
